Modularize save event
